requirements.txt
python-dotenv
schedule
groq
requests